Carry the raw int32 exit code through process.reallyExit
Process_functionReallyExit masked process.exit(N)'s code to a u8 before
handing it to Bun__Process__exit, so in a worker the parent's
worker.on("exit", code) saw 44 for process.exit(300) while the
process.exitCode = 300 spelling correctly produced 300. Node passes the
raw int32 to the parent for both. The main thread is unaffected: the OS
status is still masked to 0-255 at the final exit(2).

Also drain the subprocess stderr pipes in the new tests, replace the
setTimeout in the rejection-routing fixture with setImmediate, and pin
the int32 wrap boundary (process.exitCode = 2**31 reads back as
-2147483648, matching Node).

+++ src/jsc/VirtualMachine.rs
@@ -1715,7 +1715,7 @@ pub struct RuntimeHooks {
     /// it returns and the caller `panic!`s. Lives in `bun_runtime::node`
     /// (forward-dep cycle), so [`uncaught_exception`] reaches it through this
     /// slot instead of the linker.
-    pub process_exit: unsafe fn(global: *mut JSGlobalObject, code: u8),
+    pub process_exit: unsafe fn(global: *mut JSGlobalObject, code: i32),
     /// `node_cluster_binding.handleInternalMessageChild(global, data)`.
     pub handle_ipc_internal_child: unsafe fn(global: *mut JSGlobalObject, data: JSValue),
     /// `node_cluster_binding.child_singleton.deinit()`.
